feat: introduce core module for basic lit features
