+++ packages/core/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "@lit-protocol/core",
+  "version": "2.2.5",
+  "type": "commonjs",
+  "license": "MIT",
+  "homepage": "https://github.com/Lit-Protocol/js-sdk",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Lit-Protocol/js-sdk"
+  },
+  "keywords": [
+    "library"
+  ],
+  "bugs": {
+    "url": "https://github.com/Lit-Protocol/js-sdk/issues"
+  },
+  "publishConfig": {
+    "access": "public",
+    "directory": "../../dist/packages/core"
+  },
+  "browser": {
+    "crypto": false,
+    "stream": false
+  },
+  "tags": [
+    "universal"
+  ],
+  "main": "./dist/src/index.js",
+  "typings": "./dist/src/index.d.ts"
+}
